Modify ComboBox and make it reusable component.
