+++ components/HomePage/ComboBox.tsx
@@ -1,10 +1,15 @@
 'use client';
 
-import * as React from 'react';
+import React from 'react';
 import { CaretSortIcon, CheckIcon } from '@radix-ui/react-icons';
 
 import { cn } from '@/lib/utils';
 import { Button } from '@/components/ui/button';
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/ui/popover';
 import {
   Command,
   CommandEmpty,
@@ -13,28 +18,10 @@ import {
   CommandItem,
   CommandList,
 } from '@/components/ui/command';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover';
 
-const boards = [
-  {
-    value: 'Job Search 2024',
-    label: 'Job Search 2024',
-  },
-  {
-    value: 'Job Search 2023',
-    label: 'Job Search 2023',
-  },
-  {
-    value: 'Job Search 2022',
-    label: 'Job Search 2022',
-  },
-];
+type ComboBoxInput = { value: string; label: string };
 
-export function ComboBox() {
+export function ComboBox({ items }: { items: ComboBoxInput[] }) {
   const [open, setOpen] = React.useState(false);
   const [value, setValue] = React.useState('');
 
@@ -48,21 +35,21 @@ export function ComboBox() {
           className="w-[200px] justify-between"
         >
           {value
-            ? boards.find((board) => board.value === value)?.label
+            ? items.find((item) => item.value === value)?.label
             : 'Job Search 2024'}
           <CaretSortIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
         </Button>
       </PopoverTrigger>
       <PopoverContent className="w-[200px] p-0">
         <Command>
-          <CommandInput placeholder="Search board..." className="h-9" />
+          <CommandInput placeholder="Search item..." className="h-9" />
           <CommandList>
-            <CommandEmpty>No board found.</CommandEmpty>
+            <CommandEmpty>No item found.</CommandEmpty>
             <CommandGroup>
-              {boards.map((board) => (
+              {items.map((item) => (
                 <CommandItem
-                  key={board.value}
-                  value={board.value}
+                  key={item.value}
+                  value={item.value}
                   onSelect={(currentValue) => {
                     setValue(currentValue === value ? '' : currentValue);
                     setOpen(false);
@@ -71,10 +58,10 @@ export function ComboBox() {
                   <CheckIcon
                     className={cn(
                       'mr-2 h-4 w-4',
-                      value === board.value ? 'opacity-100' : 'opacity-0'
+                      value === item.value ? 'opacity-100' : 'opacity-0'
                     )}
                   />
-                  {board.label}
+                  {item.label}
                 </CommandItem>
               ))}
             </CommandGroup>
