+++ data/boards.ts
@@ -1,0 +1,16 @@
+const boards = [
+  {
+    value: 'Job Search 2024',
+    label: 'Job Search 2024',
+  },
+  {
+    value: 'Job Search 2023',
+    label: 'Job Search 2023',
+  },
+  {
+    value: 'Job Search 2022',
+    label: 'Job Search 2022',
+  },
+];
+
+export default boards;
